requirements.txt
aiohttp==3.9.3
azure-identity==1.18.0
azure-search-documents==11.6.0b4
python-dotenv==1.0.1
azure-storage-blob==12.23.1
gunicorn
rich
openai==1.58.1
azure-cosmos==4.9.0